Fix healthcheck + stale SW + legacy earnings reset
- Add /api/health endpoint (bypasses auth) for Railway healthcheck
- Bump SW cache lila-v4 → lila-v5 to force CSS re-fetch on clients
- Cast numeric comparison explicitly to clear legacy \$1247.50 seed

+++ lib/db.ts
@@ -27,8 +27,8 @@ export async function ensureSchema(client: PoolClient): Promise<void> {
       updated_at      TIMESTAMPTZ   NOT NULL DEFAULT NOW()
     );
     INSERT INTO lila_state (id) VALUES (1) ON CONFLICT DO NOTHING;
-    -- Reset legacy seed data
-    UPDATE lila_state SET total_earned = 0 WHERE id = 1 AND total_earned = 1247.50;
+    -- Wipe legacy seed value — only touches the exact fake amount, never real earnings
+    UPDATE lila_state SET total_earned = 0 WHERE id = 1 AND total_earned::numeric = 1247.50::numeric;
     -- Add assigned_bounty column if upgrading from earlier schema
     ALTER TABLE lila_state ADD COLUMN IF NOT EXISTS assigned_bounty JSONB DEFAULT NULL;
